Update Rockpaperscissors.py
Games are now only recorded if the deep model is being used. This is because players will probably play a lot more differently if they are playing against a program that simply chooses a move randomly, i.e they may choose randomly themselves or keep choosing the same option. If the model is to be effective, it would ideally be trained on data from players choosing more rationally.